[UPDATE] rename tables->collections for mongo coherency
